+++ src/server/overlay/OverlayHub.cpp
@@ -46,7 +46,13 @@ void OverlayHub::configure(void *arg) {
 		auto netmaskStr = conf.getString("OVERLAY", "netMask", "0x0");
 		sscanf(netmaskStr, "%llx", &ctx.netMask);
 		ctx.groupId = conf.getNumber("OVERLAY", "groupId");
-		loadIdentifiers("BOOTSTRAP", "nodes", ctx.bootstrapNodes, 128);
+
+		if (!Identity::getIdentifiers("BOOTSTRAP", "nodes", ctx.bootstrapNodes,
+				WH_ARRAYLEN(ctx.bootstrapNodes))) {
+			auto n = Identity::getIdentifiers(ctx.bootstrapNodes,
+			WH_ARRAYLEN(ctx.bootstrapNodes) - 1, Host::BOOTSTRAP);
+			ctx.bootstrapNodes[n] = 0;
+		}
 
 		WH_LOG_DEBUG(
 				"Overlay hub settings: \n" "ENABLE_REGISTRATION=%s, AUTHENTICATE_CLIENTS=%s, CONNECT_TO_OVERLAY=%s,\n" "TABLE_UPDATE_CYCLE=%ums, BLOCKING_IO_TIMEOUT=%ums, RETRY_INTERVAL=%ums,\n" "NETMASK=%s, GROUP_ID=%u\n",