+++ src/server/overlay/OverlayService.cpp
@@ -28,7 +28,7 @@ OverlayService::~OverlayService() {
 
 void OverlayService::setBootstrapNodes(const unsigned long long *nodes) noexcept {
 	unsigned int i = 0;
-	for (; nodes && nodes[i] && (i < 15); ++i) {
+	for (; nodes && nodes[i] && (i < (WH_ARRAYLEN(ctx.nodes) - 1)); ++i) {
 		ctx.nodes[i] = nodes[i];
 	}