+++ docs/ChangeLog.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Const versions of the **Message::getHeader** and **Message::getStorage** methods.
+- **Message::getHeader** and **Message::getStorage** methods with const modifier.
+- Bootstrap using the hosts database or file.
 
 ### Changed
 
@@ -17,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Hub::removeWatcher** returns boolean value.
 - **Inotifier::read** doesn't unnecessarily zero out the IO buffer.
 - Add const modifier to the method parameters in the **Endpoint** and **Protocol** classes (wherever applicable).
+- Rename **Identity::loadIdentifiers** to **Identity::getIdentifiers**.
 
 ## [3.0.0] - 2021-03-02